Pass the instance to ctor instead of class table
__call gets the class table as its first argument, so passing those args
to obj.new meant we had the wrong self in the constructor. Since the
class is our __index, when we failed to find member variables in the
instance we found them in the class.

Also add test case that exposed this error: the new channel would get
printed because the self captured in new was the class table.